backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
dynaconf==3.2.4
python-multipart==0.0.9
httpx==0.27.0
PyJWT==2.8.0
web3==6.20.0
aioboto3==13.1.1
aiosqlite==0.20.0